# Benchmarks and Improvements for parseRequestURL function