--- a/middleware_test.go
+++ b/middleware_test.go
@@ -106,6 +106,46 @@ func Test_parseRequestURL(t *testing.T) {
 			expectedURL: "https://example.com/4%2F5/6/7",
 		},
 		{
+			name: "empty path parameter in URL",
+			init: func(c *Client, r *Request) {
+				r.SetPathParams(map[string]string{
+					"bar": "4",
+				})
+				r.URL = "https://example.com/{}/{bar}"
+			},
+			expectedURL: "https://example.com/%7B%7D/4",
+		},
+		{
+			name: "not closed path parameter in URL",
+			init: func(c *Client, r *Request) {
+				r.SetPathParams(map[string]string{
+					"foo": "4",
+				})
+				r.URL = "https://example.com/{foo}/{bar/1"
+			},
+			expectedURL: "https://example.com/4/%7Bbar/1",
+		},
+		{
+			name: "extra path parameter in URL",
+			init: func(c *Client, r *Request) {
+				r.SetPathParams(map[string]string{
+					"foo": "1",
+				})
+				r.URL = "https://example.com/{foo}/{bar}"
+			},
+			expectedURL: "https://example.com/1/%7Bbar%7D",
+		},
+		{
+			name: " path parameter with remainder",
+			init: func(c *Client, r *Request) {
+				r.SetPathParams(map[string]string{
+					"foo": "1",
+				})
+				r.URL = "https://example.com/{foo}/2"
+			},
+			expectedURL: "https://example.com/1/2",
+		},
+		{
 			name: "using BaseURL with absolute URL in request",
 			init: func(c *Client, r *Request) {
 				c.SetBaseURL("https://foo.bar") // ignored