The benchmarks for the applying PathParams and adding QueryParams.

Original results:
```shell
% go test -benchmem -bench=. -run=^Benchmark
goos: darwin
goarch: amd64
pkg: github.com/go-resty/resty/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_parseRequestURL_PathParams-16    524658    2260 ns/op    448 B/op     9 allocs/op
Benchmark_parseRequestURL_QueryParams-16   865923    1371 ns/op    416 B/op    13 allocs/op
```

After the improvements:
```shell
% go test -benchmem -bench=. -run=^Benchmark
goos: darwin
goarch: amd64
pkg: github.com/go-resty/resty/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
Benchmark_parseRequestURL_PathParams-16     753834    1367 ns/op    256 B/op    5 allocs/op
Benchmark_parseRequestURL_QueryParams-16   1000000    1167 ns/op    352 B/op    9 allocs/op
```

The applying of PathParams has been improved by using O(1) logic instead of O(N). (N the number of PathParams + RawPathParams in both client and request). Instead of calling `strings.Replace` for each path parameter, the current logic collects all needed parameters in a map then replaces all parameters in URL by searching for the curly brackets and replacing. It scans the whole URL just once, from first to last positions.

The improvements of adding the QueryParams have been done by swapping the processing order. First - process the request's QueryParams, only then process the client's QueryParams and skip already existed instead of deleting.