@SVilgelm 2023-09-29
```
% benchstat old.txt new.txt
goos: darwin
goarch: amd64
pkg: github.com/go-resty/resty/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
                                │   old.txt    │               new.txt                │
                                │    sec/op    │    sec/op     vs base                │
_parseRequestHeader-16            191.9n ± ∞ ¹   190.7n ± ∞ ¹       ~ (p=1.000 n=1) ²
_parseRequestURL_PathParams-16    1.368µ ± ∞ ¹   1.321µ ± ∞ ¹       ~ (p=1.000 n=1) ²
_parseRequestURL_QueryParams-16   1.414µ ± ∞ ¹   1.130µ ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                           718.7n         657.8n        -8.47%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                │   old.txt   │               new.txt                │
                                │    B/op     │    B/op      vs base                 │
_parseRequestHeader-16            0.000 ± ∞ ¹   0.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestURL_PathParams-16    320.0 ± ∞ ¹   256.0 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestURL_QueryParams-16   416.0 ± ∞ ¹   352.0 ± ∞ ¹        ~ (p=1.000 n=1) ³
geomean                                     ⁴                -12.20%               ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ need >= 4 samples to detect a difference at alpha level 0.05
⁴ summaries must be >0 to compute geomean

                                │   old.txt    │               new.txt                │
                                │  allocs/op   │  allocs/op   vs base                 │
_parseRequestHeader-16             0.000 ± ∞ ¹   0.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
_parseRequestURL_PathParams-16     6.000 ± ∞ ¹   5.000 ± ∞ ¹        ~ (p=1.000 n=1) ³
_parseRequestURL_QueryParams-16   13.000 ± ∞ ¹   9.000 ± ∞ ¹        ~ (p=1.000 n=1) ³
geomean                                      ⁴                -16.75%               ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ need >= 4 samples to detect a difference at alpha level 0.05
⁴ summaries must be >0 to compute geomean
```