@SVilgelm 2023-10-02
New benchmarks, almost not changes
```
goos: darwin
goarch: amd64
pkg: github.com/go-resty/resty/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
                                │   old.txt    │               new.txt               │
                                │    sec/op    │   sec/op     vs base                │
_parseRequestHeader-16             189.1n ± 3%   188.7n ± 2%        ~ (p=0.361 n=10)
_parseRequestURL_PathParams-16     1.564µ ± 0%   1.304µ ± 1%  -16.62% (p=0.000 n=10)
_parseRequestURL_QueryParams-16   1347.5n ± 1%   971.0n ± 2%  -27.94% (p=0.000 n=10)
geomean                            735.8n        620.5n       -15.68%

                                │   old.txt    │               new.txt                │
                                │     B/op     │    B/op     vs base                  │
_parseRequestHeader-16            0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
_parseRequestURL_PathParams-16    448.0 ± 0%     256.0 ± 0%  -42.86% (p=0.000 n=10)
_parseRequestURL_QueryParams-16   416.0 ± 0%     352.0 ± 0%  -15.38% (p=0.000 n=10)
geomean                                      ²               -21.51%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                │    old.txt    │               new.txt                │
                                │   allocs/op   │ allocs/op   vs base                  │
_parseRequestHeader-16             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
_parseRequestURL_PathParams-16     9.000 ± 0%     5.000 ± 0%  -44.44% (p=0.000 n=10)
_parseRequestURL_QueryParams-16   13.000 ± 0%     9.000 ± 0%  -30.77% (p=0.000 n=10)
geomean                                       ²               -27.28%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```