@SVilgelm 2023-10-02
```
goos: darwin
goarch: amd64
pkg: github.com/go-resty/resty/v2
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
                                │   old.txt    │               new.txt               │
                                │    sec/op    │   sec/op     vs base                │
_parseRequestURL_PathParams-16     1.570µ ± 1%   1.269µ ± 1%  -19.15% (p=0.000 n=10)
_parseRequestURL_QueryParams-16   1343.5n ± 2%   999.3n ± 4%  -25.62% (p=0.000 n=10)
geomean                            1.452µ        1.126µ       -22.45%

                                │  old.txt   │              new.txt               │
                                │    B/op    │    B/op     vs base                │
_parseRequestURL_PathParams-16    448.0 ± 0%   256.0 ± 0%  -42.86% (p=0.000 n=10)
_parseRequestURL_QueryParams-16   416.0 ± 0%   352.0 ± 0%  -15.38% (p=0.000 n=10)
geomean                           431.7        300.2       -30.46%

                                │   old.txt   │              new.txt               │
                                │  allocs/op  │ allocs/op   vs base                │
_parseRequestURL_PathParams-16     9.000 ± 0%   5.000 ± 0%  -44.44% (p=0.000 n=10)
_parseRequestURL_QueryParams-16   13.000 ± 0%   9.000 ± 0%  -30.77% (p=0.000 n=10)
geomean                            10.82        6.708       -37.98%
```